Move refactored ChildrenDetailsView to PITH_views folder.
