+++ application/views/PITH_views/PITH_children_details.py
@@ -1,57 +1,49 @@
+from datetime import date
+
 from django.http import HttpResponseRedirect
+from django.shortcuts import render, reverse
+from django.utils import timezone
+from django.views.generic import View
 
-from application.utils import get_id
-from application.views.PITH_views.base_views.PITH_radio_view import PITHRadioView
-from application.forms.PITH_forms.PITH_adult_details_form import PITHAdultDetailsForm
+from application.business_logic import (
+    other_people_children_details_logic,
+    rearrange_children_in_home,
+    remove_child_in_home,
+    reset_declaration,
+)
+from application.forms import OtherPeopleChildrenDetailsForm
+from application.models import Application
 
-from application.business_logic import remove_adult, rearrange_adults, get_application
 
-
-class PITHChildrenDetailsView(PITHRadioView):
-    template_name = 'PITH_templates/PITH_adults_details.html'
-    form_class = PITHAdultDetailsForm
-    success_url = ('PITH-Adult-Details-View', 'PITH-Children-Check-View')
-    application_field_name = 'adults_in_home'
-
-    def get(self, request, *args, **kwargs):
-        application_id = get_id(request)
-        num_adults = int(request.GET.get('adults'))
-        remove_person = int(request.GET.get('remove'))
-
-        if remove_person:
-            # Remove adult flagged for being removed.
-            remove_adult(application_id, remove_person)
-
-        # Rearrange adults to remove empty spaces in adult list.
-        rearrange_adults(num_adults, application_id)
-
+class PITHChildrenDetailsView(View):
+    """
+    Class containing the methods responsible for handling requests to the 'Children-In-The-Home-Details' page.
+    """
+    def get(self, request):
         application_id_local = request.GET["id"]
+        application = Application.objects.get(pk=application_id_local)
+
         number_of_children = int(request.GET["children"])
         remove_person = int(request.GET["remove"])
         remove_button = True
-        # If there are no adults in the database
-        if number_of_children == 0:
-            # Set the number of children to 1 to initialise one instance of the form
-            number_of_children = 1
-            # Disable the remove person button
-            remove_button = False
-        # If there is only one child in the database
+
+        if number_of_children == 0:  # If there are no children in the database
+            number_of_children = 1   # Set the number of children to 1 to initialise one instance of the form
+
         if number_of_children == 1:
-            # Disable the remove person button
-            remove_button = False
-        application = Application.objects.get(pk=application_id_local)
+            remove_button = False    # Disable the remove person button
+
         remove_child_in_home(application_id_local, remove_person)
         rearrange_children_in_home(number_of_children, application_id_local)
-        # Generate a list of forms to iterate through in the HTML
-        form_list = []
-        for i in range(1, number_of_children + 1):
-            form = OtherPeopleChildrenDetailsForm(
-                id=application_id_local, child=i, prefix=i)
-            if application.application_status == 'FURTHER_INFORMATION':
+
+        form_list = [OtherPeopleChildrenDetailsForm(id=application_id_local, child=i, prefix=i) for i in range(1, number_of_children + 1)]
+
+        if application.application_status == 'FURTHER_INFORMATION':
+            for index, form in enumerate(form_list):
                 form.error_summary_template_name = 'returned-error-summary.html'
-                form.error_summary_title = "There was a problem (Child " + str(i) + ")"
-            form_list.append(form)
-            form.check_flag()
+                form.error_summary_title = "There was a problem (Child " + str(index + 1) + ")"
+                form.check_flag()
+
         variables = {
             'form_list': form_list,
             'application_id': application_id_local,
@@ -62,61 +54,115 @@ class PITHChildrenDetailsView(PITHRadioView):
             'people_in_home_status': application.people_in_home_status
         }
 
-        return super().get(request, *args, **kwargs)
+        return render(request, 'other-people-children-details.html', variables)
 
-    def get_context_data(self, **kwargs):
-        application_id = get_id(self.request)
-        num_adults = int(self.request.GET.get('adults'))
-        PITH_status = get_application(application_id, 'people_in_home_status')
+    def post(self, request):
+        current_date = timezone.now()
+        application_id_local = request.POST["id"]
+        application = Application.objects.get(pk=application_id_local)
 
-        remove_person = int(self.request.GET.get('remove'))
+        number_of_children = int(request.POST["children"])
+        remove_button = True
 
-        # If no adults exist, initialise one instance of the form by setting to var to 1.
-        num_adults = 1 if num_adults == 0 else num_adults
+        if number_of_children == 0:  # If there are no children in the database
+            number_of_children = 1   # Set the number of children to 1 to initialise one instance of the form
 
-        # Disable the remove button if number_of_adults is 1.
-        show_remove_button = not num_adults == 1
+        if number_of_children == 1:
+            remove_button = False  # Disable the remove person button
 
-        email_list = self.get_email_list(num_adults)
+        form_list   = []
+        forms_valid = True           # Bool indicating whether or not all the forms are valid
+        children_turning_16 = False  # Bool indicating whether or not all any children are turning 16
 
-        form_list = [self.create_form_instance(application_id, index, email_list) for index in range(1, num_adults + 1)]
+        for i in range(1, int(number_of_children) + 1):
+            form = OtherPeopleChildrenDetailsForm(request.POST, id=application_id_local, child=i, prefix=i)
+            form.remove_flag()
+            form.error_summary_title = 'There was a problem with the details (Child ' + str(i) + ')'
+            form_list.append(form)
 
-        context = {
-            'form': None,
-            'form_list': form_list,
-            'application_id': application_id,
-            'number_of_adults': num_adults,
-            'add_adult': num_adults + 1,
-            'remove_adult': num_adults - 1,
-            'remove_button': show_remove_button,
-            'people_in_home_status': PITH_status
-        }
+            if application.application_status == 'FURTHER_INFORMATION':
+                form.error_summary_template_name = 'returned-error-summary.html'
 
-        return context
+            if form.is_valid():
+                child_record = other_people_children_details_logic(application_id_local, form, i)
+                child_record.save()
+                reset_declaration(application)
 
-    def create_form_instance(self, app_id, adult_index, email_list):
-        form_class = self.get_form_class()
+                # Calculate child's age
+                birth_day, birth_month, birth_year = form.cleaned_data.get('date_of_birth')
+                applicant_dob = date(birth_year, birth_month, birth_day)
+                today = date.today()
 
-        new_form = form_class(id=app_id,
-                              adult=adult_index,
-                              prefix=adult_index,
-                              email_list=email_list)
-        new_form.check_flag()
+                age = today.year - applicant_dob.year - ((today.month, today.day) < (applicant_dob.month, applicant_dob.day))
+                if 15 <= age < 16:
+                    children_turning_16 = True
 
-        return new_form
+            else:
+                forms_valid = False
 
-    def get_email_list(self, num_adults):
-        return ['' for index in range(1, num_adults + 1)]
+        if 'submit' in request.POST:
+            # If all forms are valid
+            if forms_valid:
+                variables = {
+                    'application_id': application_id_local,
+                    'people_in_home_status': application.people_in_home_status,
+                }
 
-    def form_valid(self, form):
+                if children_turning_16:
+                    application.children_turning_16 = True
+                    success_url = 'Other-People-Approaching-16-View' # If a child is approaching 16, navigate to approaching 16 page
+                else:
+                    application.children_turning_16 = False
+                    success_url = 'Other-People-Summary-View' # If no child is approaching 16, navigate to summary page
 
-        context = {
-            'id': get_id(self.request),
-            'adults': 0,
-            'remove': 0
-        }
+                application.date_updated = current_date
+                application.save()
+                reset_declaration(application)
+                return HttpResponseRedirect(reverse(success_url) + '?id=' + application_id_local, variables)
 
-        return HttpResponseRedirect(self.get_success_url(get=context))
+            # If there is an invalid form
+            else:
+                variables = {
+                    'form_list': form_list,
+                    'application_id': application_id_local,
+                    'number_of_children': number_of_children,
+                    'add_child': int(number_of_children) + 1,
+                    'remove_child': int(number_of_children) - 1,
+                    'remove_button': remove_button,
+                    'people_in_home_status': application.people_in_home_status
+                }
+                return render(request, 'other-people-children-details.html', variables)
+
+        if 'add_child' in request.POST:
+            # If all forms are valid
+            if forms_valid:
+                variables = {
+                    'application_id': application_id_local,
+                    'people_in_home_status': application.people_in_home_status
+                }
+
+                add_child = int(number_of_children) + 1
+                add_child_string = str(add_child)
+
+                # Redirect to self.get(), it seems.
+                return HttpResponseRedirect(reverse('PITH-Children-Details-View') + \
+                                            '?id=' + application_id_local + \
+                                            '&children=' + add_child_string + \
+                                            '&remove=0#person' + add_child_string,
+                                            variables)
+
+            # If there is an invalid form
+            else:
+                variables = {
+                    'form_list': form_list,
+                    'application_id': application_id_local,
+                    'number_of_children': number_of_children,
+                    'add_child': int(number_of_children) + 1,
+                    'remove_child': int(number_of_children) - 1,
+                    'remove_button': remove_button,
+                    'people_in_home_status': application.people_in_home_status
+                }
+                return render(request, 'other-people-children-details.html', variables)
