+++ application/views/other_people.py
@@ -9,12 +9,11 @@ import pytz
 from django.utils import timezone
 
 import calendar
-from datetime import date, datetime, timedelta
+from datetime import datetime
 
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist
 from django.http import HttpResponseRedirect
-from django.views.generic import View
 from django.shortcuts import render, reverse
 
 from application.utils import build_url
@@ -25,18 +24,13 @@ from ..summary_page_data import other_adult_link_dict, other_adult_name_dict, ot
     other_child_summary_name_dict, other_child_summary_link_dict
 from ..business_logic import (health_check_email_resend_logic,
                               other_people_adult_details_logic,
-                              other_people_children_details_logic,
                               rearrange_adults,
-                              rearrange_children_in_home,
                               remove_adult,
-                              remove_child_in_home,
                               reset_declaration,
                               show_resend_and_change_email)
 from ..forms import (
                      OtherPeopleAdultDetailsForm,
                      OtherPeopleApproaching16Form,
-                     OtherPeopleChildrenDetailsForm,
-                     OtherPeopleChildrenQuestionForm,
                      OtherPeopleEmailConfirmationForm,
                      OtherPeopleResendEmailForm,
                      OtherPeopleSummaryForm)
@@ -206,156 +200,6 @@ def other_people_adult_details(request):
                     'people_in_home_status': application.people_in_home_status
                 }
                 return render(request, 'other-people-adult-details.html', variables)
-
-
-class OtherPeopleChildrenDetailsView(View):
-    """
-    Class containing the methods responsible for handling requests to the 'Children-In-The-Home-Details' page.
-    """
-    def get(self, request):
-        application_id_local = request.GET["id"]
-        application = Application.objects.get(pk=application_id_local)
-
-        number_of_children = int(request.GET["children"])
-        remove_person = int(request.GET["remove"])
-        remove_button = True
-
-        if number_of_children == 0:  # If there are no children in the database
-            number_of_children = 1   # Set the number of children to 1 to initialise one instance of the form
-
-        if number_of_children == 1:
-            remove_button = False    # Disable the remove person button
-
-        remove_child_in_home(application_id_local, remove_person)
-        rearrange_children_in_home(number_of_children, application_id_local)
-
-        form_list = [OtherPeopleChildrenDetailsForm(id=application_id_local, child=i, prefix=i) for i in range(1, number_of_children + 1)]
-
-        if application.application_status == 'FURTHER_INFORMATION':
-            for index, form in enumerate(form_list):
-                form.error_summary_template_name = 'returned-error-summary.html'
-                form.error_summary_title = "There was a problem (Child " + str(index + 1) + ")"
-                form.check_flag()
-
-        variables = {
-            'form_list': form_list,
-            'application_id': application_id_local,
-            'number_of_children': number_of_children,
-            'add_child': number_of_children + 1,
-            'remove_button': remove_button,
-            'remove_child': number_of_children - 1,
-            'people_in_home_status': application.people_in_home_status
-        }
-
-        return render(request, 'other-people-children-details.html', variables)
-
-    def post(self, request):
-        current_date = timezone.now()
-        application_id_local = request.POST["id"]
-        application = Application.objects.get(pk=application_id_local)
-
-        number_of_children = int(request.POST["children"])
-        remove_button = True
-
-        if number_of_children == 0:  # If there are no children in the database
-            number_of_children = 1   # Set the number of children to 1 to initialise one instance of the form
-
-        if number_of_children == 1:
-            remove_button = False  # Disable the remove person button
-
-        form_list   = []
-        forms_valid = True           # Bool indicating whether or not all the forms are valid
-        children_turning_16 = False  # Bool indicating whether or not all any children are turning 16
-
-        for i in range(1, int(number_of_children) + 1):
-            form = OtherPeopleChildrenDetailsForm(request.POST, id=application_id_local, child=i, prefix=i)
-            form.remove_flag()
-            form.error_summary_title = 'There was a problem with the details (Child ' + str(i) + ')'
-            form_list.append(form)
-
-            if application.application_status == 'FURTHER_INFORMATION':
-                form.error_summary_template_name = 'returned-error-summary.html'
-
-            if form.is_valid():
-                child_record = other_people_children_details_logic(application_id_local, form, i)
-                child_record.save()
-                reset_declaration(application)
-
-                # Calculate child's age
-                birth_day, birth_month, birth_year = form.cleaned_data.get('date_of_birth')
-                applicant_dob = date(birth_year, birth_month, birth_day)
-                today = date.today()
-
-                age = today.year - applicant_dob.year - ((today.month, today.day) < (applicant_dob.month, applicant_dob.day))
-                if 15 <= age < 16:
-                    children_turning_16 = True
-
-            else:
-                forms_valid = False
-
-        if 'submit' in request.POST:
-            # If all forms are valid
-            if forms_valid:
-                variables = {
-                    'application_id': application_id_local,
-                    'people_in_home_status': application.people_in_home_status,
-                }
-
-                if children_turning_16:
-                    application.children_turning_16 = True
-                    success_url = 'Other-People-Approaching-16-View' # If a child is approaching 16, navigate to approaching 16 page
-                else:
-                    application.children_turning_16 = False
-                    success_url = 'Other-People-Summary-View' # If no child is approaching 16, navigate to summary page
-
-                application.date_updated = current_date
-                application.save()
-                reset_declaration(application)
-                return HttpResponseRedirect(reverse(success_url) + '?id=' + application_id_local, variables)
-
-            # If there is an invalid form
-            else:
-                variables = {
-                    'form_list': form_list,
-                    'application_id': application_id_local,
-                    'number_of_children': number_of_children,
-                    'add_child': int(number_of_children) + 1,
-                    'remove_child': int(number_of_children) - 1,
-                    'remove_button': remove_button,
-                    'people_in_home_status': application.people_in_home_status
-                }
-                return render(request, 'other-people-children-details.html', variables)
-
-        if 'add_child' in request.POST:
-            # If all forms are valid
-            if forms_valid:
-                variables = {
-                    'application_id': application_id_local,
-                    'people_in_home_status': application.people_in_home_status
-                }
-
-                add_child = int(number_of_children) + 1
-                add_child_string = str(add_child)
-
-                # Redirect to self.get(), it seems.
-                return HttpResponseRedirect(reverse('PITH-Children-Details-View') + \
-                                            '?id=' + application_id_local + \
-                                            '&children=' + add_child_string + \
-                                            '&remove=0#person' + add_child_string,
-                                            variables)
-
-            # If there is an invalid form
-            else:
-                variables = {
-                    'form_list': form_list,
-                    'application_id': application_id_local,
-                    'number_of_children': number_of_children,
-                    'add_child': int(number_of_children) + 1,
-                    'remove_child': int(number_of_children) - 1,
-                    'remove_button': remove_button,
-                    'people_in_home_status': application.people_in_home_status
-                }
-                return render(request, 'other-people-children-details.html', variables)
 
 
 def other_people_approaching_16(request):
