+++ childminder/urls.py
@@ -114,7 +114,7 @@ urlpatterns = [
 
     # Children
     url(r'^people/children/$', PITH_views.PITHChildrenCheckView.as_view(), name='PITH-Children-Check-View'),
-    url(r'^people/children-details/', views.OtherPeopleChildrenDetailsView.as_view(),
+    url(r'^people/children-details/', PITH_views.PITHChildrenDetailsView.as_view(),
         name='PITH-Children-Details-View'),
     url(r'^people/children-turning-16/', views.other_people_approaching_16, name='Other-People-Approaching-16-View'),
     url(r'^people/check-answers/', views.other_people_summary, name='Other-People-Summary-View'),
